// src/app/models/survey.ts
export class Survey {
    user: string | undefined | null;
    name: string | undefined;
    surname: string | undefined;
    age: number | undefined;
    phone: string | undefined;
    foundBy: string | undefined;
    liked: number | undefined;
    favGames = [];
    recommended: string | undefined;
}